Ab upgrades

See merge request minetasia/idarkay/minetasiacore!23

+++ api/src/main/java/fr/idarkay/minetasia/core/api/utils/MinetasiaPlayer.java
@@ -148,7 +148,7 @@ public interface MinetasiaPlayer
     /**
      * @return a core value
      */
-    int getStatus();
+    long getStatus();
 
     /**
      * @return the party of the player or null if player isn"t in party
